Dockerfile just setup prerequisite now.
Normal workflow of development is
1. change js file
2. reload in browser and check
3. commit
So it's better share editbook files between host and container.

In this change, we first create image that contains all the prerequisite,
then mount editbook folder when docker run, and build editbook inside that container.
build_and_exec.sh is utility shell script for this purpose.