fix the infinite loop when langserver exits.
Sometimes the langserver binary can start successfully but
exit very quickly, and in that case editbook wrongly read
empty messages repeatedly.

The primary bug is the handling of scanner. As described
in https://golang.org/pkg/bufio/#Scanner.Scan, when a scanner
reached to the end of the stream (which happens when the process
exits), scanner.Err() will still return nil. In that case
ReadMessage() should return io.EOF actually.

This patch also includes some feature to print out errors
when the process exits unexpectedly.